Replace tomllib with json
Currently the ArchAdept example project config files are in TOML format
which requires the `tomllib` library in Python, which was added to
standard Python in Python 3.11.

Rather than requiring such a new Python for so little benefit, let's
simply replace the config files with equivalent JSON files. JSON has
been part of the Python standard library for a very long time so we
can keep the Python required version down at 3.8.