Remove `-D` flag from `archadept make dis`
Now that we've switched over to the LLVM toolchain we're using
`llvm-objdump` instead of GNU `objdump`, and `llvm-objdump`'s `-D`
flag does *not* disassemble data found in text sections (i.e. literal
pools and explicit use of directives like `.word`). That was the only
reason we allowed for the `-D` flag to be passed through, so remove it.