[bug] Fix crash when rendering errors
Commit d3a967604f971 introduced picking up the user's color palette
when rendering Rich text and panels, instead of hardcoding the color.

However, a missing `style=` when rendering errors meant that we were
erroneously passing the value of `Color.RED` -- that is, `'color(1)'`,
through to the Rich panel as a positional arg rather than a kwarg,
which crashes the Rich renderer because raw strings do not implement
the Rich renderable API.

While we're here, ensure we always render errors even when `-v` is
passed on the command line to enable debug logging.